+++ src/pages/forgot-password/index.tsx
@@ -1,6 +1,6 @@
 import { useRouter } from 'next/router';
 import toast from 'react-hot-toast';
-import { useForgotPassword } from '@/features/account/hooks/useForgotPassword';
+import { useForgotPassword } from '@/features/account/hooks/use-forgot-password';
 import { Meta } from '@/components/meta';
 import { Container } from '@/components/container';
 import { Title } from '@/components/title';
