+++ src/pages/my-account/index.tsx
@@ -6,8 +6,8 @@ import toast from 'react-hot-toast';
 import {
     getProfile,
     useGetProfile
-} from '@/features/manage/hooks/useGetProfile';
-import { useDeleteAccount } from '@/features/manage/hooks/useDeleteAccount';
+} from '@/features/manage/hooks/use-get-profile';
+import { useDeleteAccount } from '@/features/manage/hooks/use-delete-account';
 import { authOptions } from '@/pages/api/auth/[...nextauth]';
 import { Meta } from '@/components/meta';
 import { Container } from '@/components/container';
