+++ src/pages/register/index.tsx
@@ -1,7 +1,7 @@
 import { useRouter } from 'next/router';
 import { signIn } from 'next-auth/react';
 import toast from 'react-hot-toast';
-import { useRegister } from '@/features/account/hooks/useRegister';
+import { useRegister } from '@/features/account/hooks/use-register';
 import { Meta } from '@/components/meta';
 import { Container } from '@/components/container';
 import { Title } from '@/components/title';
