+++ src/pages/user/[id]/index.tsx
@@ -3,11 +3,11 @@ import { getServerSession } from 'next-auth';
 import { useRouter } from 'next/router';
 import { dehydrate, QueryClient } from '@tanstack/react-query';
 import toast from 'react-hot-toast';
-import { getUser, useGetUser } from '@/features/user/hooks/useGetUser';
-import { useUpdateUser } from '@/features/user/hooks/useUpdateUser';
-import { useLockUser } from '@/features/user/hooks/useLockUser';
-import { useUnlockUser } from '@/features/user/hooks/useUnlockUser';
-import { useDeleteUser } from '@/features/user/hooks/useDeleteUser';
+import { getUser, useGetUser } from '@/features/user/hooks/use-get-user';
+import { useUpdateUser } from '@/features/user/hooks/use-update-user';
+import { useLockUser } from '@/features/user/hooks/use-lock-user';
+import { useUnlockUser } from '@/features/user/hooks/use-unlock-user';
+import { useDeleteUser } from '@/features/user/hooks/use-delete-user';
 import { authOptions } from '@/pages/api/auth/[...nextauth]';
 import { Meta } from '@/components/meta';
 import { Container } from '@/components/container';
